languagemonster/requirements.txt
amqp==1.4.6
anyjson==0.3.3
astroid==1.4.4
Babel==2.2.0
backports.ssl-match-hostname==3.5.0.1
betterpath==0.2.2
billiard==3.3.0.20
biplist==0.9
celery==3.1.18
certifi==2016.2.28
click==6.3
colorama==0.3.7
coverage==4.0.3
defusedxml==0.4.1
Django==1.9.4
django-appconf==1.0.1
django-celery==3.1.17
django-celery-email==1.1.2
django-countries==3.3
django-debug-toolbar==1.4
django-filter==0.9.2
django-mongonaut==0.2.20
django-nose==1.4.3
django-post-office==2.0.6
django-profiler==2.0.0
django-redis-cache==1.6.5
django-redisboard==3.0.0
django-rest==0.0.1
django-rest-swagger==0.3.5
django-rosetta==0.7.10
django-system-monitor==0.5
django-tastypie==0.12.1
djangorestframework==3.3.2
factory-boy==2.6.1
fake-factory==0.5.7
flake8==2.5.4
flower==0.9.0
funcsigs==0.4
futures==3.0.5
ipaddress==1.0.16
isort==4.2.2
jsonfield==1.0.3
kombu==3.0.26
l==0.9.1
lazy-object-proxy==1.2.1
Markdown==2.6
mccabe==0.4.0
meld3==1.0.2
microsofttranslator==0.7
mock==1.3.0
nose==1.3.7
oauthlib==0.7.2
pbr==1.8.1
pep8==1.7.0
Pillow==2.7.0
polib==1.0.7
psutil==1.2.1
psycopg2==2.6.1
py==1.4.31
pyflakes==1.0.0
PyJWT==1.3.0
pylint==1.5.4
pytest==2.9.1
python-dateutil==2.4.1
python-memcached==1.57
python-mimeparse==0.1.4
python-openid==2.2.5
python-social-auth==0.2.9
pytz==2015.4
PyYAML==3.11
redis==2.10.3
requests==2.9.1
requests-oauthlib==0.5.0
selenium==2.53.1
six==1.10.0
sorl-thumbnail==12.3
sqlparse==0.1.19
supervisor==3.1.3
tornado==4.2
vcversioner==2.14.0.0
wrapt==1.10.6
zope.interface==4.1.3
